obs-ffmpeg: Fix comparison type mismatch

+++ plugins/obs-ffmpeg/jim-nvenc.c
@@ -597,7 +597,7 @@ static bool init_encoder(struct nvenc_data *enc, obs_data_t *settings,
 static bool init_bitstreams(struct nvenc_data *enc)
 {
 	da_reserve(enc->bitstreams, enc->buf_count);
-	for (size_t i = 0; i < enc->buf_count; i++) {
+	for (int i = 0; i < enc->buf_count; i++) {
 		struct nv_bitstream bitstream;
 		if (!nv_bitstream_init(enc, &bitstream)) {
 			return false;
@@ -612,7 +612,7 @@ static bool init_bitstreams(struct nvenc_data *enc)
 static bool init_textures(struct nvenc_data *enc)
 {
 	da_reserve(enc->bitstreams, enc->buf_count);
-	for (size_t i = 0; i < enc->buf_count; i++) {
+	for (int i = 0; i < enc->buf_count; i++) {
 		struct nv_texture texture;
 		if (!nv_texture_init(enc, &texture)) {
 			return false;
